Rewrite card filtering for performance
Use event dispatchers instead of interfaces
Refactor & cleanup

+++ Source/DeckBuilderDesktop/DataTables/FCardData.h
@@ -19,7 +19,7 @@ public:
 		FString CardName;
 
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = Card)
-		float Cost;
+		int32 Cost;
 
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = Card)
 		FString Type;
